+++ x/evm/core/vm/custom_eip_test.go
@@ -147,3 +147,97 @@ func TestExtendActivators(t *testing.T) {
 		tc.postCheck()
 	}
 }
+
+func TestAddOperation(t *testing.T) {
+	// Functions used to create an operation.
+	customExecute := func(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext) ([]byte, error) {
+		// no - op
+		return nil, nil
+	}
+	customDynamicGas := func(evm *EVM, contract *Contract, stack *Stack, mem *Memory, memorySize uint64) (uint64, error) {
+		// no-op
+		return 0, nil
+	}
+	customMemorySize := func(stack *Stack) (uint64, bool) {
+		// no-op
+		return 0, false
+	}
+
+	const (
+		EXISTENT OpCode = STOP
+		NEW      OpCode = 0xf
+	)
+
+	testCases := []struct {
+		name        string
+		opName      string
+		opNumber    OpCode
+		expPass     bool
+		errContains string
+		postCheck   func()
+	}{
+		{
+			"fail - operation with same number already exists",
+			"TEST",
+			EXISTENT,
+			false,
+			"already exists",
+			func() {
+				name := EXISTENT.String()
+				require.Equal(t, "STOP", name)
+			},
+		},
+		{
+			"fail - operation with same name already exists",
+			"CREATE",
+			NEW,
+			false,
+			"already exists",
+			func() {
+				name := NEW.String()
+				require.Contains(t, name, "not defined")
+			},
+		},
+		{
+			"fail - operation with same name of STOP",
+			"STOP",
+			NEW,
+			false,
+			"already exists",
+			func() {
+				name := NEW.String()
+				require.Contains(t, name, "not defined")
+			},
+		},
+		{
+			"pass - new operation added to the list",
+			"TEST",
+			NEW,
+			true,
+			"",
+			func() {
+				name := NEW.String()
+				require.Equal(t, "TEST", name)
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			opInfo := OpCodeInfo{
+				Number: tc.opNumber,
+				Name:   tc.opName,
+			}
+			_, err := ExtendOperations(opInfo, customExecute, 0, customDynamicGas, 0, 0, customMemorySize)
+
+			if tc.expPass {
+				require.NoError(t, err)
+			} else {
+				require.Error(t, err)
+				require.Contains(t, err.Error(), tc.errContains, "expected different error")
+			}
+
+			tc.postCheck()
+		})
+	}
+}